controllers,internal: Move the reporting related functions to their own dedicated package
